+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `sandstone/Scroller` with `editable` prop to move item via 5-way keys in pointer mode
+
 ## [2.9.0] - 2024-07-17
 
 ### Added