+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ### Fixed
 
-- `sandstone/Steps` prop `size` to accept Number type
+- `sandstone/Steps` prop `size` to accept number type
 
 ## [2.9.0] - 2024-07-17