+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 - `sandstone/Alert` public class names `alert`, `content`, `fullscreen`, and `title`
 
+### Changed
+
+- `sandstone/Icon` supported icon list, adding new icons
+
 ## [2.7.17] - 2024-07-08
 
 ### Added